refactor(gfs-ext): split the 1768-line lib.rs into focused modules
The planner_hook extension lived in a single src/lib.rs. Split it by concern
into route / keyrange / pushdown / federate / catalog / hydrate / model
modules (lib.rs keeps the crate root + the planner hook), and move the
catalog DDL out to src/sql/schema.sql loaded via extension_sql_file!.

Behaviour-preserving: the SQL is byte-identical and the extension compiles
clean; verified end-to-end by rebuilding the gfs-postgres:16 image from these
modules and running the benchmark — identical results (14/14, same metrics).
